+++ webpack.config.js
@@ -1,0 +1,85 @@
+const path = require('path');
+
+const { VueLoaderPlugin } = require('vue-loader');
+const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+const webpack = require('webpack');
+
+const buildEnv = process.argv.includes('--mode=production')
+  ? 'production'
+  : 'development';
+const devMode = buildEnv === 'development';
+const jsPrefix = 'src/main/resources/frontend';
+
+function entrypointPath(filename) {
+  return path.join(__dirname, jsPrefix, filename);
+}
+
+module.exports = {
+  mode: buildEnv,
+  externals: {
+    jquery: 'jQuery'
+  },
+  entry: {
+    'delivery-details': entrypointPath('delivery-details.js')
+  },
+  resolve: {
+    alias: {
+      '@': path.join(__dirname, jsPrefix)
+    },
+    extensions: ['.js', '.vue', '.json']
+  },
+  output: {
+    path: path.join(__dirname, 'target/classes/static/assets'),
+    filename: 'js/[name].js'
+  },
+  module: {
+    rules: [
+      {
+        test: /\.vue$/,
+        loader: 'vue-loader'
+      },
+      {
+        test: /\.js$/,
+        exclude: /node_modules/,
+        loader: 'babel-loader',
+        options: {
+          presets: ['@babel/preset-env']
+        }
+      },
+      {
+        test: /\.css$/,
+        use: [devMode ? 'style-loader' : MiniCssExtractPlugin.loader, 'css-loader']
+      },
+      {
+        test: /\.(png|svg|jpg|gif)$/,
+        loader: 'file-loader',
+        options: {
+          name: 'assets/img/[name].[ext]'
+        }
+      }
+    ]
+  },
+  devtool: 'cheap-source-map',
+  plugins: [
+    new MiniCssExtractPlugin({
+      filename: 'css/[name].css'
+    }),
+    new VueLoaderPlugin(),
+    new webpack.DefinePlugin({
+      __VUE_OPTIONS_API__: JSON.stringify(true),
+      __VUE_PROD_DEVTOOLS__: JSON.stringify(devMode),
+      __VUE_PROD_HYDRATION_MISMATCH_DETAILS__: JSON.stringify(false)
+    })
+  ],
+  optimization: {
+    splitChunks: {
+      cacheGroups: {
+        vendor: {
+          test: /[\\/]node_modules[\\/]/,
+          name: 'notificationlib-vendor',
+          chunks: 'initial'
+        }
+      }
+    }
+  }
+};